requirements.txt
feedparser
httpx
rich
pandas
rapidfuzz
pyyaml
python-dotenv
py_clob_client
web3
